% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTBLMM.R
\name{RunTBLMM}
\alias{RunTBLMM}
\title{Run TBLMM and make prediction}
\usage{
RunTBLMM(
  phenofile,
  gene_path = NULL,
  exp_path = NULL,
  meth_path = NULL,
  kernel = NULL,
  readFun = readRDS,
  verbose = TRUE,
  modelFit = FALSE
)
}
\arguments{
\item{phenofile}{path to the phenotype files that have 3 columns (PID, IID and Y). The individuals to be predicted should be marked as \code{NA}.}

\item{gene_path}{path to each genotype file (current implementation expects to have genotype files).}

\item{exp_path}{path to each gene expression file.}

\item{meth_path}{path to each methylation file.}

\item{kernel}{kernel function for each omics (default is linear kernel.)}

\item{readFun}{the function used to read gene expression, genotype and methylation files. The default is \code{readRDS}.}

\item{verbose}{print the fitting process (default is \code{TRUE}).}

\item{modelFit}{return all model fits including parameter estimates (default is \code{FALSE}, and only predicted values to be exported).}
}
\value{
A list of predicted values.
}
\description{
Run TBLMM and make prediction
}
\examples{
datadir= system.file("extdata", package = "TBLMM")
gene_path=paste0(datadir,"/geno",1:2,".RDS")
expr_path=paste0(datadir,"/expr",1:2,".RDS")
meth_path=paste0(datadir,"/meth",1:2,".RDS")
phenofile=paste0(datadir,"/pheno.txt")
test=RunTBLMM(phenofile=phenofile,gene_path=gene_path,exp_path=expr_path, meth_path=meth_path,
kernel=c("linear","linear","linear"),readFun=readRDS)
}
